nature reserve created

+++ 09_OOP1/practice/src/main/kotlin/Bird.kt
@@ -1,10 +1,10 @@
-class Bird(energy: Int, weight: Double, currentAge: Int, maxAge: Int, name: String) :
+class Bird(energy: Int, weight: Int, currentAge: Int, maxAge: Int, name: String) :
     Animal(energy, weight, currentAge, maxAge, name) {
     override fun move(message: String) {
         super.move("$name flying")
     }
 
-    override fun giveOffspring(energy: Int, weight: Double, currentAge: Int, maxAge: Int, name: String): Bird {
+    override fun giveOffspring(energy: Int, weight: Int, currentAge: Int, maxAge: Int, name: String): Bird {
         super.giveOffspring(energy, weight, currentAge, maxAge, name)
         return Bird(energy, weight, currentAge, maxAge, name)
     }
