+++ 09_OOP1/practice/src/main/kotlin/NatureReserve.kt
@@ -1,10 +1,10 @@
 open class NatureReserve {
-    fun getBird(energy: Int, weight: Double, currentAge: Int, maxAge: Int, name: String) =
+    fun getBird(energy: Int, weight: Int, currentAge: Int, maxAge: Int, name: String) =
         Bird(energy, weight,currentAge, maxAge, name)
-    fun getFish(energy: Int, weight: Double, currentAge: Int, maxAge: Int, name: String) =
+    fun getFish(energy: Int, weight: Int, currentAge: Int, maxAge: Int, name: String) =
         Fish(energy, weight,currentAge, maxAge, name)
-    fun getDog(energy: Int, weight: Double, currentAge: Int, maxAge: Int, name: String) =
+    fun getDog(energy: Int, weight: Int, currentAge: Int, maxAge: Int, name: String) =
         Dog(energy, weight,currentAge, maxAge, name)
-    fun getAnimal(energy: Int, weight: Double, currentAge: Int, maxAge: Int, name: String) =
+    fun getAnimal(energy: Int, weight: Int, currentAge: Int, maxAge: Int, name: String) =
         Animal(energy, weight,currentAge, maxAge, name)
 }
